Split Placeholder into a separate file
